+++ src/main/resources/application.properties
@@ -15,7 +15,7 @@ spring.thymeleaf.mode=HTML
 spring.thymeleaf.encoding=UTF-8
 spring.thymeleaf.cache=false
 spring.devtools.livereload.enabled=true
-spring.devtools.restart.enabled=true
+spring.devtools.restart.enabled=false
